Fixes 'User selects to create User Feed' test in `FeedEditModal.spec.ts` + more
- Fixed broken 'User selects to create User Feed' test in `FeedEditModal.spec.ts`. There are other tests in the same file that still must be fixed.
- Created a new file to hold route details - `/src/lib/router.ts`.
